Rewrite the webapp to use core gem
This include a lot of changes, I'm sorry I didn't split this into several
commits, but now it would take too much time to split it.
Anyway, changes are simple:
- Rename some paths: lib/model -> lib/models.
- Rename some paths: lib/util -> lib/utils.
- Rewrite some utils as Sinatra's helpers.
- Use $LOAD_PATH as it's meant to be used, instead of cumbersome
requires.
- Update every use access to the environment to use "the new way".
- Update every access to Redis to use the new way. Notice, as a very
nice side effect of this refactoring that Redis is access at a single
point in the whole app (in `lib/helpers/redis.rb`).