Finished refactoring of all models
This finishes a complete review of the app at the model level. The
biggest change is the use of STI for Teacher/Students, etc. The models
were identical and the db columns too, so it seems like a good fit (at
least for now) and it simplifies the code.

Also, the reporting methods has been reworked and moved to
`Page.summary` and `Site.summary`. The common functionality has been
extracted to a `Summarizable` module.