Implement page visit tracking using a heartbeat mechanism
The previous tracking system that used browser events to detect when the
page was unloaded or the window/tab closed is not valid in all cases, so
we have decided to implement a heartbeat based solution that sends a
periodic 'viewed' message to the server with the elapsed time.

Some details about the implementation:

* No new raw messages will be imported as long as one of them exists
that is unprocessed and has the same heartbeat id; in this case we just
update its current time and 'captured_at' value. This ensures that only
the most recent message, and time, is considered.
* When the extraction of page visits is executed, it will look for the
latest raw message and create the visit with the same time +
heartbeat id. If a visit already exists for the same heartbeat id it
will mean that the user has not yet abandoned the page, so the visit
record will be updated instead.
* A custom tracking interval can be specified by setting the
'tracking_interval' parameter when calling the collector.js or loader.js
scripts.