Perform better detection of current window status
It uses the Page Visibility API to catch some new scenarios and treat
them correctly. However, this API only detects, for the time being, a
limited set of actions but not all of them (i.e.: changing the window or
alt-tabbing to another application does not raise the event), so the
time spent on a page can never be 100% accurate.