Load visualization data via API
This way we can reload it via JS on demand, for example when the user
changes the date range.

Also, moved the visualizer backend into a separate module. Now the
visualizer code is largely self contained, in /lib/visualizer, so it
should be easy to extract it into a separate application later if
necessary.